Use 'ftpsubdir' variable (set to 'releases/' or 'snapshots/' depending on the build 'type') to stage the FTP bits.
This creates an additional subdirectory structure to the 'ftpdir', so for
now, comment out the echo output for rsync(1) at the end; this should also
be conditionalized.

This is intended to allow running a full build of all architectures (snap
and release types), without needing further separation between the ftp
target staging directories.  Additionally, this allows automation of
staging the vm images correctly, regardless of what the target builds are.



git-svn-id: file:///Users/gjb/Downloads/releng-svn/thermite-r688@18 5711f056-dcf5-4cc1-a278-612e92bcda7e