Code reduction for the chroot build phases.
With how the filesystem is organized for the chroot builds,
there is no longer a need to separate the builds for head/
from the builds for stable/ branches.

This was a relic of original (disorganized) logic that would
map the base/${branch} directly to the filesystem, instead
of using the revision number directly; there no longer is an
'extra' directory in the layout.



git-svn-id: file:///Users/gjb/Downloads/releng-svn/thermite-r688@32 5711f056-dcf5-4cc1-a278-612e92bcda7e