Fix ftp-stage.sh for new arm configuration layout.
Export FTP_STAGING=1 in ftp-stage.  When defined and the
build configuration is sourced, call load_stage_env(), which
calls load_target_env() and then sets TARGET and TARGET_ARCH
to the XDEV and XDEV_ARCH.

While here, fix a few exports in the PANDABOARD configuration
that were missed in the last update.

Sponsored by:	The FreeBSD Foundation


git-svn-id: file:///Users/gjb/Downloads/releng-svn/thermite-r688@264 5711f056-dcf5-4cc1-a278-612e92bcda7e